model: adjust data type

+++ src/main/java/restful/api/eztrain/model/UpdateRouteRequest.java
@@ -24,6 +24,6 @@ public class UpdateRouteRequest {
     
     private Double tripDistance;
     
-    private Integer tripDuration;
+    private Double tripDuration;
 
 }
